requirements.txt
# Core Libraries
Pillow==11.0.0           # For image processing
opencv-python==4.8.0.74 # For advanced image manipulation (deskewing, cropping)

# PDF Generation with OC
fpdf==1.7.2             # To create PDFs from images
PyPDF2
pdf2image               # Optional, for pre-existing PDFs

# OCR (Optional)
pytesseract    # For OCR (requires Tesseract installed separately)

# File Watching (Optional)
watchdog==3.0.0         # For real-time monitoring of directories (optional)

# Utilities (Optional)
rich==13.6.0            # For better CLI output (optional)

#GUI
ttkbootstrap